wip: Work on exposing scoped D-Bus services
